Module Bap_c

C language support library.

This library brings support for C Abstract Machine. In particular it adds a fairly complete support for C type system, data model, and ABI.

open Bap_c.Std module to use this library, it only defines one module C that includes the rest of the library. It also defines an interface for parser, that maybe provided by a third party.

module Std : sig ... end