Module Bap_c_size

An abstraction of sizeof operator.

type bits = Core_kernel.Int.t

val next_multitude_of : n:int -> int -> int

next_multitude_of ~n x returns y >= x so that y is a multitude of n, i.e., y = n * k.

val padding : Bap.Std.Size.t -> int -> int

padding alignment offset computes the required padding at offset to ensure the alignment.

val max_enum_elt : (string * int64 option) list -> int64

max_enum_elt enum returns the maximum element in the enum specification.

class base : Bap_c_data.model -> object ... end

The base class for computing sizes and aligments of C data types.