Module Bap_c_size

type bits = Core_kernel.Int.t

class base : Bap_c_data.model -> object ... end

Base class for computing size of C data types. The algorithm is implemented as a class to allow a particular implementation to fine tune the calculation. We need here an open recursion, since type is inherently recursive.