Module Bap_c_term_attributes

val data : Bap_c_data.t Bap.Std.tag

Abstraction of a data representation of C value. This attribute is attached to each inserted arg term, but can be further propagated by other passes

val proto : Bap_c_type.proto Bap.Std.tag

Function prototype. This attribute is inserted into each annotated function.

val t : Bap_c_type.t Bap.Std.tag

A c type associated with a term. This attribute is attached to each inserted arg term, but maybe propagated by further by other passes.